Lesson 11: Authorization

+++ routers/posts.py
@@ -11,6 +11,8 @@ from sqlalchemy.orm import selectinload
 from sqlalchemy.ext.asyncio import AsyncSession
 import models
 
+from auth import CurrentUser
+
 # Pydantic Schemas imports
 from schemas import PostResponse, PostCreate, PostUpdate
 
@@ -39,14 +41,8 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
     return post
 
 @router.post('', response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User doesn't exist")
-    
-    new_post = models.Post(title=post.title, content=post.content, user_id=post.user_id)
+async def create_post(post: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
+    new_post = models.Post(title=post.title, content=post.content, user_id=current_user.id)
 
     db.add(new_post)
     await db.commit()
@@ -55,7 +51,8 @@ async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_
     return new_post
 
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id: int, current_user: CurrentUser, 
+                           post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]) -> models.Post:
     # check if the post that needs to be changed exists
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
@@ -63,16 +60,11 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[As
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
-    # check if the user requesting an update exists
-    result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
-    user = result.scalars().first()
-
-    if not user:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update the post")
     
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -80,7 +72,8 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[As
     return post
 
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id: int, post_data: PostUpdate, 
+                              current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]) -> models.Post:
     # check if the post to be updated exists
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
@@ -88,6 +81,9 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
     
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update the post")
+
     # this line excludes fields in the request which have None values
     update_data = post_data.model_dump(exclude_unset=True)
     for key, value in update_data.items():
@@ -101,12 +97,15 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
 # DELETE Methods doesn't have response body, instead it returns HTTP 204 code, which means that
 # request was successfull, and deletion took place
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete the post")
     
     await db.delete(post)
     await db.commit()
